+++ core/components/romanescobackyard/docs/changelog.md
@@ -3,6 +3,7 @@
 ## Backyard 1.0.0-beta16
 Released on ...
 
+- Include PHPStan via Composer
 - Add styling for right floated icons in MIGX grids
 - Add search field to input options grid
 - Add new input option groups through modal