+++ core/components/romanescobackyard/docs/changelog.md
@@ -3,6 +3,7 @@
 ## Backyard 1.0.0-beta16
 Released on ...
 
+- Make search field smaller in MIGX TV grids
 - Avoid horizontal scrollbars in MIGX modals with nested grid
 - Show thumbnail next to selector after selecting image
 - Replace updated combo box value with ghost element